Download Kirby21 Imaging Data

This function allows users to download specific modalities for Kirby21 data sets. This allows this package to be on CRAN

download_kirby21_data(
  modality = kirby21.base::all_modalities(),
  progress = TRUE,
  force = FALSE,
  outdir = NULL
)

Arguments

modality

modality of images that are to be downloaded. You must have the package downloaded for that modality.

progress

Should verbose messages be printed when downloading the data

force

If the package of that modality is not installed stop. If force = FALSE, then this will download the data but not really install the package.

outdir

output directory for files to download. It will default to the directory of the corresponding package for the data.

Value

A logical indicating the data is there.

Examples

on_cran = !identical(Sys.getenv("NOT_CRAN"), "true")
on_ci <- nzchar(Sys.getenv("CI"))
local_run = grepl("musch", tolower(Sys.info()[["user"]]))
run_example = !on_cran || on_ci || local_run
if (run_example) {
outdir = tempdir()
res = download_kirby21_data("SURVEY", outdir = outdir)
}
#> cloning into '/tmp/RtmpgPPKBG/file75491a99d3bb/kirby21.survey'...
#> Receiving objects:   1% (1/78),   16 kb
#> Receiving objects:  11% (9/78),   16 kb
#> Receiving objects:  21% (17/78),   16 kb
#> Receiving objects:  32% (25/78),   16 kb
#> Receiving objects:  41% (32/78),  771 kb
#> Receiving objects:  51% (40/78), 2289 kb
#> Receiving objects:  61% (48/78), 4547 kb
#> Receiving objects:  71% (56/78), 4547 kb
#> Receiving objects:  82% (64/78), 4547 kb
#> Receiving objects:  91% (71/78), 4547 kb
#> Receiving objects: 100% (78/78), 4547 kb, done.